+++ CHANGELOG.md
@@ -17,6 +17,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Handled newlines after Google argument types.  Newlines
   were handled in most other situations (inside of types,
   after an untyped item, etc.) but this one slipped through.
+- Handle parentheses inside parenthetical google types.
+  Previously, darglint simply failed to parse those arguments.
+  Now it will raise a ParameterMalformedError.
 
 ## [1.3.1]